moves analytics files into analytics folder
