+++ static/app/plugins/components/settings.tsx
@@ -7,7 +7,7 @@ import LoadingIndicator from 'app/components/loadingIndicator';
 import {t, tct} from 'app/locale';
 import {Organization, Plugin, Project} from 'app/types';
 import {parseRepo} from 'app/utils';
-import {IntegrationAnalyticsKey} from 'app/utils/integrationEvents';
+import {IntegrationAnalyticsKey} from 'app/utils/analytics/integrationAnalyticsEvents';
 import {trackIntegrationEvent} from 'app/utils/integrationUtil';
 
 type Props = {
