+++ static/app/utils/analytics/growthAnalyticsEvents.tsx
