+++ static/app/utils/analytics/issueAnalyticsEvents.tsx
