+++ static/app/views/performance/transactionDetails/finishSetupAlert.tsx
@@ -5,7 +5,7 @@ import {IconLightning} from 'app/icons';
 import {t} from 'app/locale';
 import space from 'app/styles/space';
 import {Organization, Project} from 'app/types';
-import {trackAdvancedAnalyticsEvent} from 'app/utils/advancedAnalytics';
+import {trackAdvancedAnalyticsEvent} from 'app/utils/analytics/advancedAnalytics';
 
 export default function FinishSetupAlert({
   organization,
